GH-500: validate compiled data directories
