+++ contributing/development/benchmarking-and-performance-validation.md
@@ -2,7 +2,7 @@
 
 - Preserve the fast path. Profile hot code with the `benchmarks/` pytest-benchmark suites (`make speedtest` for a quick core-subset check, `make benchmarks` for the full suite) when touching polygon math or shortcut lookups.
 - Memory is measured too, by `make memory` - see *Benchmarking* below. If a change makes the library hold data it previously mapped or streamed, that shows up there and nowhere else.
-- Validate generated artifacts where they are generated, not where they are read. The correctness of a compiled data directory is established once, by the build; re-checking it when a `TimezoneFinder` is constructed spends every user's startup time re-answering a question that already has an answer, and multiplies across the per-thread instances concurrent workloads are told to use. The check belongs in two places instead - the converter, over the files it just wrote, and the test suite, over the files the repository ships - sharing one implementation (`scripts/data_integrity.py`) so that the two cannot drift into asserting different things. The payoff is not only speed: a check that no longer has to fit in an initialisation budget can afford to be exhaustive, which is why the hole-reference check resolves every ring in the dataset and compares it against an independently derived bounding box.
+- Validate generated artifacts where they are generated, not automatically where they are read. The correctness of a compiled data directory is established once, by the build; re-checking it when a `TimezoneFinder` is constructed spends every user's startup time re-answering a question that already has an answer, and multiplies across the per-thread instances concurrent workloads are told to use. The converter checks the files it just wrote, the test suite checks the files the repository ships, and users who compile custom data can opt in with `timezonefinder validate-data`; all three share `timezonefinder/_data_integrity.py` so they cannot drift into asserting different things. The payoff is not only speed: a check that no longer has to fit in an initialisation budget can afford to be exhaustive, which is why the hole-reference check resolves every ring in the dataset and compares it against an independently derived bounding box.
 - Use vectorised/NumPy-aware operations and avoid quadratic fallbacks on large datasets. When performance optimisations add complexity, include comments that summarise the micro-optimisation.
 - Respect coordinate scaling constants and FlatBuffers layouts; keep performance-sensitive structures (H3 mappings, bbox filters) cache-friendly. When changing a FlatBuffers schema, delete any previously generated `.bin` binary artifacts so they are regenerated consistently.
