+++ contributing/improvements/checked-and-found-sound/runtime-geometry-and-data-checks.md
@@ -8,6 +8,6 @@ Do not re-raise these findings without new evidence.
 
 - **Workflows and global facade:** `.github/workflows/benchmark.yml`, `benchmark-comment.yml`, `check_data_updates.yml`, and `release_data_update.yml` were read in full and yielded nothing beyond version drift that is now guarded. `timezonefinder/global_functions.py`'s module-level `TF_INSTANCE` and `global` statement (Ruff `PLW0603`) are deliberately documented as not thread-safe, with per-thread instances as the alternative. `scripts/hex_utils.py`'s antimeridian and pole clipping in `get_corrected_hex_boundaries` is covered by `tests/hex_utils_test.py`.
 
-- **Data integrity and package boundaries:** `scripts/data_integrity.py`'s two validators each build their own `PolygonArray`/`HoleArray`, but they are separate entry points with different subjects—cross-file agreement and an upstream-data expectation—and `__del__` releases the accessors either way. `packages/timezonefinder-data/timezonefinder_data/__init__.py` and `scripts/data_releases.py` were read in full and found sound. `timezonefinder/zone_names.py` deliberately lets `read_zone_names` take a default path while requiring one for `write_zone_names`, because defaulting the write side to `DEFAULT_DATA_DIR` could rewrite the installed dataset in `site-packages`.
+- **Data integrity and package boundaries:** the general directory validators live in private `timezonefinder/_data_integrity.py` so the installed CLI can call them; `scripts/data_integrity.py` retains only the packaged-upstream deduplication expectation that would reject valid custom data. Their `PolygonArray`/`HoleArray` instances release accessors through `__del__`. `packages/timezonefinder-data/timezonefinder_data/__init__.py` and `scripts/data_releases.py` were read in full and found sound. `timezonefinder/zone_names.py` deliberately lets `read_zone_names` take a default path while requiring one for `write_zone_names`, because defaulting the write side to `DEFAULT_DATA_DIR` could rewrite the installed dataset in `site-packages`.
 
 - **A failed certainty probe does not by itself prove a shortcut omission.** `certain_timezone_at()` tests every candidate but the point-in-polygon kernel has no promised side for a point exactly on a stored boundary. At the exact South Pole, and at border offsets smaller than the package's coordinate grid, an exhaustive scan can find no containing packaged polygon either. Diagnose an index omission only when exhaustive point-in-polygon testing finds a containing polygon absent from the cell's candidates; `None` alone establishes only that no candidate contained the quantized point.
