+++ contributing/improvements/decisions/benchmarking-tooling-and-dependency-decisions.md
@@ -12,6 +12,6 @@ Do not re-propose these settled or refused options without new evidence.
 
 - **Shrink the runtime dependency surface (numpy / h3 / cffi / flatbuffers) — considered and parked.** Each does one small thing, so the idea recurs. Reimplementing H3 indexing is a well-known source of subtle bugs and `h3` sits on the common path of every query; an open item would be an invitation to attempt it. Revisit only if import time or cold start is ever measured to be a real problem.
 
-- **Dropping `flatbuffers` for the custom binary format it replaced in 6.6.0 — measured 2026-08-22 and refused.** The dependency is not what anything costs. It adds **0.033 %** to the 63 MB coordinate file, **1.1 ms** of import once numpy is loaded, and — since the coordinate accessors started resolving every polygon's offset and length once at construction — **nothing per query**, the lookup path reading coordinates with a bare `np.frombuffer`. Writing the same shortcut payload as flat vectors *inside* a FlatBuffers buffer reads in **0.050 ms against 0.039 ms** for a hand-rolled raw blob of the same bytes, so the container is worth ~11 µs and 58 bytes. What was expensive was the **shape of the old shortcut schema** — a vector of 41,162 tables decoded one by one, ~400 ms of a construction — and that was a schema question in either format, which is why replacing the schema fixed it and dropping the dependency would not have. Against ~0 gain, removing it costs the file identifiers and `layout_version` fields the load-time guards are built on, the `schemas/` copy that lets a compiled data directory be read back without the package that wrote it, the alignment guarantee the zero-copy views rest on, and the independent reference decoder that `scripts.data_integrity.validate_coordinate_offset_table` checks the fast path against. Not to be re-proposed without a measurement contradicting one of those figures.
+- **Dropping `flatbuffers` for the custom binary format it replaced in 6.6.0 — measured 2026-08-22 and refused.** The dependency is not what anything costs. It adds **0.033 %** to the 63 MB coordinate file, **1.1 ms** of import once numpy is loaded, and — since the coordinate accessors started resolving every polygon's offset and length once at construction — **nothing per query**, the lookup path reading coordinates with a bare `np.frombuffer`. Writing the same shortcut payload as flat vectors *inside* a FlatBuffers buffer reads in **0.050 ms against 0.039 ms** for a hand-rolled raw blob of the same bytes, so the container is worth ~11 µs and 58 bytes. What was expensive was the **shape of the old shortcut schema** — a vector of 41,162 tables decoded one by one, ~400 ms of a construction — and that was a schema question in either format, which is why replacing the schema fixed it and dropping the dependency would not have. Against ~0 gain, removing it costs the file identifiers and `layout_version` fields the load-time guards are built on, the `schemas/` copy that lets a compiled data directory be read back without the package that wrote it, the alignment guarantee the zero-copy views rest on, and the independent reference decoder that `timezonefinder._data_integrity.validate_payload_offset_table` checks the fast path against. Not to be re-proposed without a measurement contradicting one of those figures.
 
 ---
