+++ timezonefinder/configs.py
@@ -81,7 +81,7 @@ NO_ZONE_ID: Final[int] = -1
 # because that is the narrowest width the dataset fits: a zone id is an index into the
 # zone names, of which the packaged data has ~450 against this width's 32,767, and the
 # shortcut table already stores those same ids as int16. The bound is not left to hold
-# by luck - ``scripts.data_integrity.validate_shortcut_index`` refuses a data directory
+# by luck - ``timezonefinder._data_integrity.validate_shortcut_index`` refuses a data directory
 # whose zone count outgrows it, at build time and over the committed data, which is
 # where a width chosen by fit has to be checked. A wider dtype would double the answer
 # array for headroom no dataset can reach.
